drm/atomic: Remove frame boosting entirely
At this point, it's begining to seem as through frame boost has
outlived it's usefulness. Currently, it's only being used to raise
the CPU's minimum frequency to 748MHz while rendering.

It's no longer necessary to maintain a seperate boost level for
rendering even in the context of SchedTune. Further more, I think
it would be more viable to bump the minimum frequency to 748MHz
globally and simply allow for scaling down to hardware minimum when
the display is off. This should preserve all of the benefits of
frame boost and then some, while mitigating any negative impact
such a frequency bump may have on idle drain. By taking this
alternative approach, we can substantially reduce the complexity of
this driver while eliminating the constant bouncing between 576 and
748MHz associated with the current frame boost implementation,
which could potentially improve efficiency as well as performance.

Change-Id: I6f490e17e5cb5b00fe7273b49941b439d9a913df
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>